ai/Curricula/pyproject.toml
[project]
name = "scheduler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.3.27",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
[tool.setuptools]
packages = ["agents", "core"]
